Move type-related passes into the types sub folder
Just a re-organization.


stack-info: PR: https://github.com/bpftrace/bpftrace/pull/5019, branch: user/jordalgo/type_reorganize/1
