+++ src/ast/passes/types/cast_creator.cpp
@@ -1,4 +1,4 @@
-#include "ast/passes/cast_creator.h"
+#include "ast/passes/types/cast_creator.h"
 #include "ast/ast.h"
 #include "ast/passes/map_sugar.h"
 #include "bpftrace.h"
