+++ src/ast/passes/types/type_checker.cpp
@@ -7,11 +7,11 @@
 #include "ast/async_event_types.h"
 #include "ast/context.h"
 #include "ast/integer_types.h"
-#include "ast/passes/type_checker.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/collect_nodes.h"
+#include "ast/passes/types/type_checker.h"
+#include "ast/passes/types/type_system.h"
 #include "bpftrace.h"
 #include "btf/compat.h"
-#include "collect_nodes.h"
 #include "config.h"
 #include "log.h"
 #include "types.h"
