+++ src/ast/passes/types/type_resolver.cpp
@@ -1,19 +1,19 @@
-#include "ast/passes/type_resolver.h"
+#include "ast/passes/types/type_resolver.h"
 #include "ast/ast.h"
 #include "ast/async_event_types.h"
-#include "ast/passes/ast_transformer.h"
-#include "ast/passes/cast_creator.h"
 #include "ast/passes/clang_parser.h"
+#include "ast/passes/collect_nodes.h"
 #include "ast/passes/fold_literals.h"
 #include "ast/passes/macro_expansion.h"
 #include "ast/passes/map_sugar.h"
 #include "ast/passes/named_param.h"
-#include "ast/passes/type_applicator.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/ast_transformer.h"
+#include "ast/passes/types/cast_creator.h"
+#include "ast/passes/types/type_applicator.h"
+#include "ast/passes/types/type_system.h"
 #include "ast/visitor.h"
 #include "bpftrace.h"
 #include "btf/compat.h"
-#include "collect_nodes.h"
 #include "config.h"
 #include "config_parser.h"
 #include "log.h"
