+++ src/main.cpp
@@ -28,13 +28,13 @@
 #include "ast/passes/parser.h"
 #include "ast/passes/pid_filter_pass.h"
 #include "ast/passes/portability_analyser.h"
-#include "ast/passes/pre_type_check.h"
 #include "ast/passes/printer.h"
 #include "ast/passes/recursion_check.h"
 #include "ast/passes/resource_analyser.h"
-#include "ast/passes/type_checker.h"
-#include "ast/passes/type_resolver.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/pre_type_check.h"
+#include "ast/passes/types/type_checker.h"
+#include "ast/passes/types/type_resolver.h"
+#include "ast/passes/types/type_system.h"
 #include "benchmark.h"
 #include "bpffeature.h"
 #include "bpftrace.h"
