+++ tests/bpftrace.cpp
@@ -12,8 +12,8 @@
 #include "ast/passes/macro_expansion.h"
 #include "ast/passes/map_sugar.h"
 #include "ast/passes/named_param.h"
-#include "ast/passes/type_checker.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/type_checker.h"
+#include "ast/passes/types/type_system.h"
 #include "bpfmap.h"
 #include "bpftrace.h"
 #include "btf_common.h"
