+++ tests/control_flow_analyser.cpp
@@ -1,6 +1,6 @@
 #include "ast/passes/control_flow_analyser.h"
 #include "ast/passes/parser.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/type_system.h"
 #include "mocks.h"
 #include "gtest/gtest.h"
