+++ tests/portability_analyser.cpp
@@ -6,9 +6,9 @@
 #include "ast/passes/macro_expansion.h"
 #include "ast/passes/map_sugar.h"
 #include "ast/passes/named_param.h"
-#include "ast/passes/type_checker.h"
-#include "ast/passes/type_resolver.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/type_checker.h"
+#include "ast/passes/types/type_resolver.h"
+#include "ast/passes/types/type_system.h"
 #include "btf_common.h"
 #include "driver.h"
 #include "mocks.h"
