+++ tests/probe.cpp
@@ -9,8 +9,8 @@
 #include "ast/passes/named_param.h"
 #include "ast/passes/resolve_imports.h"
 #include "ast/passes/resource_analyser.h"
-#include "ast/passes/type_checker.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/type_checker.h"
+#include "ast/passes/types/type_system.h"
 #include "bpftrace.h"
 #include "btf.h"
 #include "btf_common.h"
