+++ tests/resource_analyser.cpp
@@ -1,8 +1,8 @@
 #include "ast/passes/resource_analyser.h"
 #include "ast/passes/parser.h"
-#include "ast/passes/type_checker.h"
-#include "ast/passes/type_resolver.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/type_checker.h"
+#include "ast/passes/types/type_resolver.h"
+#include "ast/passes/types/type_system.h"
 #include "mocks.h"
 #include "gtest/gtest.h"
