+++ tests/type_checker.cpp
@@ -18,9 +18,9 @@
 #include "ast/passes/map_sugar.h"
 #include "ast/passes/named_param.h"
 #include "ast/passes/resolve_imports.h"
-#include "ast/passes/type_checker.h"
-#include "ast/passes/type_resolver.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/type_checker.h"
+#include "ast/passes/types/type_resolver.h"
+#include "ast/passes/types/type_system.h"
 #include "ast_matchers.h"
 #include "bpftrace.h"
 #include "btf_common.h"
