+++ tests/type_resolver.cpp
@@ -9,8 +9,8 @@
 #include "ast/passes/macro_expansion.h"
 #include "ast/passes/map_sugar.h"
 #include "ast/passes/named_param.h"
-#include "ast/passes/type_resolver.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/type_resolver.h"
+#include "ast/passes/types/type_system.h"
 #include "bpftrace.h"
 #include "driver.h"
 #include "mocks.h"
