+++ tests/type_system.cpp
@@ -2,7 +2,7 @@
 
 #include "ast/ast.h"
 #include "ast/passes/clang_build.h"
-#include "ast/passes/type_system.h"
+#include "ast/passes/types/type_system.h"
 #include "mocks.h"
 
 namespace bpftrace::test::type_system {
